requirements.txt
Flask==3.0.2
feedparser==6.0.11
google-cloud-storage==2.18.2
pandas==2.2.3